+++ src/pivuq/__init__.py
@@ -1,6 +1,5 @@
 from . import disparity  # noqa: F401
 from . import lib  # noqa: F401
-from . import utils  # noqa: F401
 from .__about__ import __version__
 from .warping import warp  # noqa: F401
